Duplicated Yet More Towers, and Tweaked Radius
I've duplicated towers from the Tier1Tower, which solved the problem of
the higher up towers not hitting anything. Now I've encountered a
problem of range. No matter how high I set it (it's literally 10,000 for
the Tier2Tower), it only hits once just beyond the tiles immediately
surrounding it.